[ANI-34-auth-page-refactor] fix: fsd 구조 수정
