+++ apps/web/src/entities/auth/model/index.ts
@@ -1,0 +1,2 @@
+export { loginModalSchema, type LoginModalFormData } from './loginModalSchema'
+export { signupModalSchema, type SignupModalFormData } from './signupModalSchema'
