+++ apps/web/src/features/auth/api/login.ts
@@ -1,13 +1,13 @@
-import type { SignupRequestType, AuthResponse } from '@/__mock__/types/auth'
-import { END_POINT } from '@/shared/config/constants/end-point'
+import type { LoginRequestType, AuthResponse } from '@/__mock__/types/auth'
 import { HTTP_HEADERS, HTTP_METHODS } from '@/shared/config/constants/http'
+import { END_POINT } from '../config/auth-config'
 
-export async function signup(signupData: SignupRequestType) {
+export async function login(loginData: LoginRequestType) {
   try {
-    const response = await fetch(process.env.NEXT_PUBLIC_API_URL + END_POINT.AUTH.SIGNUP, {
+    const response = await fetch(process.env.NEXT_PUBLIC_API_URL + END_POINT.AUTH.LOGIN, {
       method: HTTP_METHODS.POST,
       headers: HTTP_HEADERS.JSON,
-      body: JSON.stringify(signupData)
+      body: JSON.stringify(loginData)
     })
 
     console.log('response', response)
@@ -19,7 +19,7 @@ export async function signup(signupData: SignupRequestType) {
     const data: AuthResponse = await response.json()
     return data
   } catch (error) {
-    console.error('회원가입 에러', error)
+    console.error('로그인 에러', error)
     return null
   }
 }
