+++ apps/web/src/features/auth/config/auth-config.ts
@@ -1,0 +1,13 @@
+export const END_POINT = {
+  AUTH: {
+    LOGIN: '/auth/login',
+    SIGNUP: '/auth/signup'
+  }
+}
+
+export const QUERY_KEYS = {
+  AUTH: {
+    LOGIN: ['authLogin'] as string[],
+    SIGNUP: ['authSignup'] as string[]
+  }
+}
