+++ apps/web/src/features/auth/model/index.ts
@@ -1,1 +1,2 @@
 export { useAuthModalStore } from './use-auth-modal-store'
+export { useAuth } from './use-auth'
