+++ apps/web/src/features/auth/model/use-auth.ts
@@ -1,0 +1,5 @@
+export function useAuth() {
+  return {
+    // React Query 쿼리 로직
+  }
+}
